feat: support string interpolating cluster ingress to build spec.setup.routeDomain
If running on OpenShift, ${CLUSTER_DOMAIN} expands to the ingress of the cluster
Make sure that the routeDomain you pick is available on the cluster.
For example, there should be NO other route on the cluster (in any namespace) with
the same domain.

For an example, check the file examples/basic.yaml

Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>